move shared UI elements to common module
